refactor: request-wrapper에 맞춰 변경
